covered SOLID and Clean that we are applying on payroll.
