+++ Kaizen/Kaizen.Server/Application/Interfaces/Payroll/IEmployeePayrollRepository.cs
@@ -1,0 +1,10 @@
+using Kaizen.Server.Application.Dtos.Payroll;
+
+namespace Kaizen.Server.Application.Interfaces.Payroll
+{
+    public interface IEmployeePayrollRepository
+    {
+        Task<List<EmployeePayroll>> GetEmployeeDataAsync(Guid companyId);
+        Task<Guid> GetPersonPkByEmailAsync(string email);
+    }
+}
