+++ Kaizen/Kaizen.Server/Application/Interfaces/Payroll/IPayrollOutputService.cs
@@ -1,0 +1,10 @@
+using Kaizen.Server.Application.Dtos.Payroll;
+
+namespace Kaizen.Server.Application.Interfaces.Payroll
+{
+    public interface IPayrollOutputService
+    {
+        void PrintPayrollSummary(PayrollSummary payrollSummary);
+        void PrintPayrollResults(List<PayrollSummary> payrollResults);
+    }
+}
