+++ Kaizen/Kaizen.Server/Application/Interfaces/Payroll/IPayrollRepository.cs
@@ -1,0 +1,9 @@
+using Kaizen.Server.Application.Dtos.Payroll;
+
+namespace Kaizen.Server.Application.Interfaces.Payroll
+{
+    public interface IPayrollRepository
+    {
+        Task SavePayrollAsync(Guid companyId, List<PayrollSummary> summaries, string email);
+    }
+}
