+++ Kaizen/Kaizen.Server/Program.cs
@@ -18,6 +18,7 @@ using Kaizen.Server.Infrastructure.Repositories.BenefitDeductions;
 using Kaizen.Server.Application.Services.ApiDeductions;
 using Kaizen.Server.Application.Services.Payroll;
 using Kaizen.Server.Application.Interfaces.Payroll;
+using Kaizen.Server.Infrastructure.Services.Payroll;
 
 
 var MyAllowSpecificOrigins = "_myAllowSpecificOrigins";
@@ -35,11 +36,11 @@ builder.Services.AddScoped<SqlConnection>(sp =>
     var connStr = config.GetConnectionString("KaizenDb");
     return new SqlConnection(connStr);
 });
-builder.Services.AddScoped<PayrollRepository>(sp =>
+builder.Services.AddScoped<Kaizen.Server.Infrastructure.Repositories.PayrollRepository>(sp =>
 {
     var config = sp.GetRequiredService<IConfiguration>();
     var connStr = config.GetConnectionString("KaizenDb");
-    return new PayrollRepository(connStr);
+    return new Kaizen.Server.Infrastructure.Repositories.PayrollRepository(connStr);
 });
 
 builder.Services.AddMemoryCache();
@@ -81,7 +82,7 @@ builder.Services.AddScoped<IDaysWorkedCalculator, DaysWorkedCalculator>();
 builder.Services.AddScoped<ISalaryCalculator, SalaryCalculator>();
 builder.Services.AddScoped<IDeductionAggregator, DeductionAggregator>();
 
-builder.Services.AddScoped <IPayrollProcessingService, PayrollProcessingService>();
+builder.Services.AddPayrollServices();
 
 builder.Services.AddMediatR(cfg =>
     cfg.RegisterServicesFromAssemblies(Assembly.GetExecutingAssembly()));
